Extract CudaProfiler from test_harness and graduate to cuda.h
